==> scripts/3_prepare_report_sections_for_labeling.ipynb <==
import os
import pandas as pd

output_dir = "/opt/gpudata/cxr-derived"

data = {
    # "chexpertplus": pd.read_csv("/opt/gpudata/chexpertplus/report.csv"),
    # "mimic": pd.read_csv("/opt/gpudata/mimic-cxr/mimic_cxr_sectioned.csv"),
    "openi": pd.read_csv("/opt/gpudata/openi/report.csv"),
}

# apply some custom filtering
for section in ["findings", "impression"]:
    for name, df in data.items():
        study_col = "study" if "study" in df else "study_id"
        temp = df[[study_col, section]].copy()
        temp[section] = temp[section].fillna("").str.strip()
        temp = temp[temp[section].str.len() > 3].reset_index(drop=True)
        temp = temp.rename(columns={study_col: "study_id"})
        temp[section].to_csv(os.path.join(output_dir, f"{name}-{section}.csv"), index=False, header=False)
        temp["study_id"].to_csv(os.path.join(output_dir, f"{name}-{section}-ids.csv"), index=False, header=True)
