==> scripts/5_cleanup_labeled_reports.ipynb <==
import os
import pandas as pd

output_dir = "/opt/gpudata/cxr-derived"
# datasets = ["chexpertplus", "mimic"]
datasets = ["openi"]
sections = ["findings", "impression"]

# apply some custom filtering
for dataset in datasets:
    for section in sections:
        for labeler in [
            # "chexpert",
            "chexbert",
        ]:
            labeled = pd.read_csv(os.path.join(output_dir, f"{dataset}-{section}-labels-{labeler}.csv"))
            ids = pd.read_csv(os.path.join(output_dir, f"{dataset}-{section}-ids.csv"))["study_id"]
            cols = labeled.columns
            labeled["study_id"] = ids
            labeled = labeled[["study_id"] + cols.to_list()]
            labeled.to_csv(os.path.join(output_dir, f"{dataset}-{section}-labels-{labeler}.csv"), index=False)
